# backend/api/mobile-v1.openapi.yaml
openapi: 3.1.0
info:
  title: FlowSpace Mobile Sync API
  version: 1.0.0
  description: >-
    Versioned contract for offline-first iPhone and Apple Watch clients.
    Existing desktop CRUD endpoints are outside this contract.
servers:
  - url: /
security:
  - sessionCookie: []
  - watchBearer: []
...
paths:
...
  /api/mobile/sync/changes:
    get:
      operationId: listMobileChanges
      summary: Read committed changes after an opaque cursor
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Scope'
        - $ref: '#/components/parameters/TimeZone'
      responses:
        '200':
          description: Change page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePage'
        '409':
          $ref: '#/components/responses/ResyncRequired'
  /api/mobile/sync/snapshot:
    get:
      operationId: getMobileSnapshot
      summary: Read an immutable paginated snapshot
      parameters:
        - $ref: '#/components/parameters/PageToken'
        - $ref: '#/components/parameters/Scope'
        - $ref: '#/components/parameters/TimeZone'
      responses:
        '200':
          description: Snapshot page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotPage'
        '410':
          $ref: '#/components/responses/ResyncRequired'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: flowspace_session
    watchBearer:
      type: http
      scheme: bearer
      bearerFormat: watch-device-token
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    PageToken:
      name: page_token
      in: query
      schema:
        type: string
    Scope:
      name: scope
      in: query
      required: true
      schema:
        type: string
        enum: [iphone, watch]
    TimeZone:
      name: time_zone
      in: query
      description: IANA time zone used by the finite Watch projection; defaults to UTC.
      schema:
        type: string
...
  responses:
...
    ResyncRequired:
      description: Cursor or snapshot session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  schemas:
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    EntityEnvelope:
      type: object
      additionalProperties: false
      required: [entity_type, entity_id, revision, payload]
      properties:
        entity_type:
          type: string
          enum: [note, task, task_occurrence, event, inbox, voice_note, transcription_job, sync_conflict]
        entity_id:
          $ref: '#/components/schemas/UUID'
        revision:
          type: integer
          minimum: 1
        deleted_at:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        payload:
          type: object
    ChangePage:
      type: object
      additionalProperties: false
      required: [schema_version, changes, next_cursor, has_more]
      properties:
        schema_version:
          const: mobile-v1
        changes:
          type: array
          items:
            $ref: '#/components/schemas/EntityEnvelope'
        next_cursor:
          type: string
        has_more:
          type: boolean
        time_zone:
          type: string
        scope_valid_until:
          $ref: '#/components/schemas/Timestamp'
        projection_refresh_required:
          type: boolean
    SnapshotPage:
      type: object
      additionalProperties: false
      required: [schema_version, entities, has_more]
      properties:
        schema_version:
          const: mobile-v1
        entities:
          type: array
          items:
            $ref: '#/components/schemas/EntityEnvelope'
        next_page_token:
          type: string
        snapshot_cursor:
          type: string
        has_more:
          type: boolean
        scope_valid_until:
          $ref: '#/components/schemas/Timestamp'
        time_zone:
          type: string
        projection_refresh_required:
          type: boolean
...
    APIError:
      type: object
      additionalProperties: false
      required: [schema_version, type, code, message]
      properties:
        schema_version:
          const: mobile-v1
        type:
          const: error
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        resync_required:
          type: boolean
...